+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added support for async modules in bridge evaluate
 * Made PHP8.1 minimum version
 * Refactored package file structure